// web/src/gen/model/agentExecutorConfigDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface AgentExecutorConfigDto {
  id: string;
  agentId: string;
  executorKey: string;
  executablePath: string;
  /** @nullable */
  version: string | null;
  createdAt: string;
}